fix: handle chunked responses in httpcore
httpcore was not properly handling response chunks. Occasionally, a response chunk did not return the full Storm message which caused a JSONDecodeError that terminated that particular Storm query. This is only common when the Storm message is exceptionally large, such as a risk:threat with long strings in the properties. It may be more common when the Cortex API is behind a proxy.

This PR writes chunked response to a buffer and relies on aiohttp.streams.StreamReader.iter_chunks to report when a chunk is "complete" before we try loading with the json library. We also catch json.JSONDecodeError exceptions within the iter_chunks loop, check if "Unterminated string starting at" is in the error, and continue the loop if so. If there is another JSON related error, it is reraised. This may have some unintended side effects when there truly is a malformed response from Synapse - but this is a bridge to cross when we actually get to it. Future work can be done to harden against this by implementing a retry limit.